requirements.txt
Django==3.2.8
django-rest-swagger==2.2.0
djangorestframework==3.12.4
psycopg2==2.9.1
flake8==4.0.1